[EC-122] refactor: Controller 분리
- attendance 도메인에 과부하 -> 분산ㄴ
