+++ api/src/main/java/org/example/educheck/domain/studentCourseAttendance/controller/StudentCourseAttendanceController.java
@@ -2,6 +2,7 @@ package org.example.educheck.domain.studentCourseAttendance.controller;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.example.educheck.domain.attendance.dto.response.MyAttendanceListResponseDto;
 import org.example.educheck.domain.member.entity.Member;
 import org.example.educheck.domain.studentCourseAttendance.dto.response.AttendanceRecordListResponseDto;
 import org.example.educheck.domain.studentCourseAttendance.service.StudentCourseAttendanceService;
@@ -40,4 +41,23 @@ public class StudentCourseAttendanceController {
                 studentCourseAttendanceService.getStudentAttendanceRecordLists(member, studentId, courseId, pageable)
         ));
     }
+
+    @PreAuthorize("hasAuthority('STUDENT')")
+    @GetMapping("/my/courses/{courseId}/attendances")
+    public ResponseEntity<ApiResponse<MyAttendanceListResponseDto>> getAttendances(@AuthenticationPrincipal Member member,
+                                                                                   @PathVariable Long courseId,
+                                                                                   @RequestParam(required = false) Integer year,
+                                                                                   @RequestParam(required = false) Integer month) {
+
+        MyAttendanceListResponseDto myAttendances = studentCourseAttendanceService.getMyAttendanceRecordLists(member, courseId, year, month);
+        studentCourseAttendanceService.getMyAttendanceRecordLists(member, courseId, year, month);
+        log.info("myAttendances: {}", myAttendances);
+
+        return ResponseEntity.ok(ApiResponse.ok(
+                "출석부 조회 성공",
+                "OK",
+                myAttendances
+        ));
+    }
+
 }
